+++ tests/testthat/helper.R
@@ -7,7 +7,7 @@ library(withr)
 ### (to my knowledge)
 skip_mkdocs <- function() {
   skip_on_cran()
-  skip_if_not(is_mkdocs())
+  skip_if_not(.is_mkdocs())
 }
 
 
@@ -60,7 +60,7 @@ create_local_thing <- function(
   old_project <- proj_get_() # this could be `NULL`, i.e. no active project
   old_wd <- getwd() # not necessarily same as `old_project`
 
-  defer(
+  withr::defer(
     {
       fs::dir_delete(dir)
     },