+++ tests/testthat/test-reformat.R
@@ -1,4 +1,4 @@
-test_that("reformat_md works", {
+test_that(".reformat_md works", {
   # Need to write and read to avoid all \n differences
   tmp <- fs::file_temp(ext = ".md")
   txt <- "# Package
@@ -25,9 +25,9 @@ Hello
 Hello again
 "
   cat(txt, file = tmp)
-  reformat_md(tmp)
-  prod <- readLines(tmp, warn = FALSE)
-  ref <- readLines(testthat::test_path("examples/examples-reformat/after-first-false.md"), warn = FALSE)
+  .reformat_md(tmp)
+  prod <- .readlines(tmp)
+  ref <- .readlines(testthat::test_path("examples/examples-reformat/after-first-false.md"))
 
   expect_identical(prod, ref)
 })
@@ -35,7 +35,7 @@ Hello again
 
 
 
-test_that("reformat_md: arg 'first' works", {
+test_that(".reformat_md: arg 'first' works", {
   # Need to write and read to avoid all \n differences
   tmp <- fs::file_temp(ext = ".md")
   txt <- "# Package
@@ -62,9 +62,9 @@ Hello
 Hello again
 "
   cat(txt, file = tmp)
-  reformat_md(tmp, first = TRUE)
-  prod <- readLines(tmp, warn = FALSE)
-  ref <- readLines(testthat::test_path("examples/examples-reformat/after-first-true.md"), warn = FALSE)
+  .reformat_md(tmp, first = TRUE)
+  prod <- .readlines(tmp)
+  ref <- .readlines(testthat::test_path("examples/examples-reformat/after-first-true.md"))
 
   expect_identical(prod, ref)
 })
@@ -72,21 +72,30 @@ Hello again
 
 
 
-test_that("replace_figures_rmd works", {
+test_that(".replace_figures_rmd works", {
+  skip_on_cran()
+  skip_if_offline()
   # setup
-  original_rmd <- readLines(
-    testthat::test_path("examples/examples-vignettes", "with-figure.Rmd"),
-    warn = FALSE
+  original_rmd <- .readlines(
+    testthat::test_path("examples/examples-vignettes", "with-figure.Rmd")
   )
   create_local_package()
-  use_docute(convert_vignettes = FALSE, path = getwd())
+  use_docute(path = getwd())
   fs::dir_create("vignettes/figures")
   fs::dir_create("docs/articles/figures")
-  download.file("https://raw.githubusercontent.com/etiennebacher/conductor/master/hex-conductor.png", "vignettes/hex-conductor.png", mode = if(.Platform$OS.type == "windows") "wb" else 'w')
-  download.file("https://raw.githubusercontent.com/etiennebacher/conductor/master/hex-conductor.png", "vignettes/figures/hex-conductor-2.png", mode = if(.Platform$OS.type == "windows") "wb" else 'w')
+  download.file(
+    "https://raw.githubusercontent.com/etiennebacher/altdoc/main/inst/misc/hex-conductor.png",
+    "vignettes/hex-conductor.png",
+    mode = if(.Platform$OS.type == "windows") "wb" else 'w'
+  )
+  download.file(
+    "https://raw.githubusercontent.com/etiennebacher/altdoc/main/inst/misc/hex-conductor.png",
+    "vignettes/figures/hex-conductor-2.png",
+    mode = if(.Platform$OS.type == "windows") "wb" else 'w'
+  )
   writeLines(original_rmd, "vignettes/with-figure.Rmd")
 
-  replace_figures_rmd()
+  .replace_figures_rmd()
   expect_true(fs::file_exists("docs/articles/figures/hex-conductor.png"))
   expect_true(fs::file_exists("docs/articles/figures/hex-conductor-2.png"))
 })